Move images to their own repository and rework deployment

+++ startup-scripts/autoexec.sh
@@ -1,4 +1,5 @@
 rm -rf genesis-client .ipython/genesis
 git clone https://github.com/dirac-institute/genesis-client
-ln -s ../genesis-client/genesis .ipython/genesis
+mkdir -p .ipython
+test -L .ipython/genesis || ln -s ../genesis-client/genesis .ipython/genesis
 gitpuller https://github.com/dirac-institute/genesis-streaming-demo master genesis-streaming-demo
